+++ NEWS.md
@@ -1,7 +1,13 @@
 # stochvol 3.0.7
 
+- NEW FUNCTION 'svlm', which has a formula interface;
+  it is a wrapper around 'svsample'
 - Turn on printing on Windows
-- Implement Geweke test in C++
+- Implement Geweke test in C++; it is feasible to
+  execute it as a CRAN test now
+- Small change in the behavior of 'predict.svdraws':
+  when 'newdata' is given then 'steps' is ignored;
+  A warning is shown if this is relevant
 
 # stochvol 3.0.6